Revert "Merge tag 'release-clm5.0.34' into TRENDY-2020"
This reverts commit 99d14b359e54423fd76a700bc587932a2df6e1ea, reversing
changes made to 54acaa1c8beb63e2d0595fa3d4ceb45c69beb672.

Go back to release-clm5.0.27

 Conflicts:
	bld/namelist_files/namelist_definition_clm4_5.xml

+++ cime_config/SystemTests/ssp.py
@@ -41,7 +41,7 @@ class SSP(SystemTestsCommon):
 
         # determine run lengths needed below
         stop_nf = self._case.get_value("STOP_N")
-        stop_n1 = int(stop_nf / 2)
+        stop_n1 = stop_nf / 2
         stop_n2 = stop_nf - stop_n1
 
         #-------------------------------------------------------------------
